moved utils.tryload to login.py
